+++ project/settings.py
@@ -1,0 +1,2 @@
+FLASK_PORT = 5002
+FLASK_DEBUG = False
